refactor(expense-documents): extract ExpenseDocumentQueryService (9 read methods, verbatim)
Phase 1 of transactional-core decompose. Behavior-preserving: 9 READ-only
methods moved verbatim to ExpenseDocumentQueryService; facade delegates,
keeps signatures. getApAging moved verbatim (aging-bucket labels untouched).
Factory wires the new service. +characterization for getApAging/getSummary/
getCreditNoteCap/getAuditTrail/previewJe. 412 existing + new tests green, tsc 0.
