+++ apps/api/src/modules/expense-documents/expense-documents.module.ts
@@ -15,6 +15,7 @@ import { DocNumberService } from './services/doc-number.service';
 import { StatusTransitionService } from './services/status-transition.service';
 import { LineAggregatorService } from './services/line-aggregator.service';
 import { JePreviewService } from './services/je-preview.service';
+import { ExpenseDocumentQueryService } from './services/expense-document-query.service';
 import { ExpenseVoucherPdfService } from './services/expense-voucher-pdf.service';
 import { PettyCashService } from './services/petty-cash.service';
 import { PayrollCustomService } from './services/payroll-custom.service';
@@ -49,6 +50,7 @@ import { ReversePermissionGuard } from './reverse-permission.guard';
     StatusTransitionService,
     LineAggregatorService,
     JePreviewService,
+    ExpenseDocumentQueryService,
     ExpenseVoucherPdfService,
     PettyCashService,
     PayrollCustomService,
